refactor: extract computeArrival into dedicated module for improved testability and maintainability
